backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-whois==0.8.0
requests==2.31.0
beautifulsoup4==4.12.2
python-Levenshtein==0.23.0
pydantic==2.5.2
dnspython==2.4.2
tldextract==3.6.0 